% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plotrow.R
\name{check_plotrow}
\alias{check_plotrow}
\title{Keep or Remove Columns}
\usage{
check_plotrow(xs = NULL, sheet = NULL)
}
\arguments{
\item{xs}{"tbl_df" "tbl" "data.frame" format from readxl::read_excel() reading an excel
spreadsheet}

\item{sheet}{a character string provided by \code{readxl::excel_sheets()} for the excel file xs}
}
\value{
\itemize{
\item \code{check_plotrow} returns a tabular format similar to xs input \verb{"tbl_df" "tbl" "data.frame"}
\item \code{check_plotrow} removes the final row \code{xs[tolower("plot"),]} if it existed, and specified
columns to not be plotted.
}
}
\description{
Check if the final row of \code{xs} contains optional "plot" filter, and apply specified filters
if so. The final row (as labelled "plot" in the "Samples" column) should contains "y" or "n"
values to dictate if the associated column should be plotted or not.
}
\examples{
# Load flowjotter's example dataset that is loaded in app by default
data(flowjotter_example)

# Ensure the first columns is titled "Samples"
flowjotter_example <- dplyr::rename(flowjotter_example, `Samples` = 1)

# Remove empty columns
flowjotter_example <- dplyr::select(flowjotter_example, dplyr::where(~ !all(is.na(.x))))

# Use check_plotrow to filter on last row.
flowjotter_example <- check_plotrow(xs = flowjotter_example, sheet = "CD4")
}
