% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_input_data.R
\name{clean_input_data}
\alias{clean_input_data}
\title{Ensure correct dataframe class conversions before plotting from input excel files.}
\usage{
clean_input_data(dirty_data = NULL)
}
\arguments{
\item{dirty_data}{input excel file from readxl::read_excel after column filtering has been
applied. First column contains groups that will be split and factored by underscore:
\code{factor(gsub("\\\\_.*", "", dirty_data[,1]))}. Other columns will contain values converted to numeric.
Other columns cannot contain ascii characters such as letters or '\%' characters.}
}
\description{
Ensures data columns are stored as numeric values, and first column \code{Samples}
is correctly split by an underscore and factorised.
}
\examples{
# Load flowjotter's example dataset that is loaded in app by default
data(flowjotter_example)

# Ensure the first columns is titled "Samples"
flowjotter_example <- dplyr::rename(flowjotter_example, `Samples` = 1)

# Remove empty columns
flowjotter_example <- dplyr::select(flowjotter_example, dplyr::where(~ !all(is.na(.x))))

# Use check_plotrow to apply filters from the last row if they exist
flowjotter_example <- check_plotrow(xs = flowjotter_example, sheet = "CD4")

# Use clean_input_data to prepare for plotting, and grouping samples.
flowjotter_example <- clean_input_data(flowjotter_example)
}
