% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pptx.R
\name{create_pptx}
\alias{create_pptx}
\title{ggplot to pptx conversion function}
\usage{
create_pptx(plot, path, left = 0.5, top = 1, width = 9, height = 4.95)
}
\arguments{
\item{plot}{ggplot2 output}

\item{path}{tempfile location to save pptx file}

\item{left}{left coordinate of the bounding box}

\item{top}{top coordinate of the bounding box}

\item{width}{width of the bounding box}

\item{height}{height of the bounding box}
}
\description{
Exports final ggplot images to single page pptx files. Could be sped up,
takes approximately 1 second per image generated. Required for pptx.
}
\examples{
# example code
create_dml <- function(plot) {
  rvg::dml(ggobj = plot)
}
g <- list()
g[[1]]    <- ggplot2::ggplot(data.frame(1),
               ggplot2::aes(x = 1, y = 1)) +
               ggplot2::geom_point()
g_dml     <- purrr::map(g, create_dml)
temp_file <- paste0(tempfile(), ".pptx")

purrr::map(
  g_dml,
  flowjotter::create_pptx,
  path = temp_file
)
}
