% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Add ability to find flowjotter image}
\usage{
.onAttach(libname, pkgname)
}
\description{
Quick fix found at
https://stackoverflow.com/questions/38791613/including-an-image-in-a-shiny-app-package
}
