% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowjotter.R
\name{flowjotter}
\alias{flowjotter}
\title{Flow Jotter Shiny App}
\usage{
flowjotter(...)
}
\description{
Use this wrapper function to load the shiny app in your RStudio Viewer, or favourite browser. No data needs to be provided to flowjotter until the shiny app is live, and data is stored locally on your machine temporarily.
}
\examples{
flowjotter()
}
