% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{flowjotter_ui}
\alias{flowjotter_ui}
\title{Shiny App UI flowjotter}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
flowjotter_ui
}
\description{
User Interface for an app for visualising Flow Cytometry data exported from Flowjo
A project started 2nd Februaryr 2023, Sam Old.
}
\keyword{datasets}
