% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plotrow.R
\name{check_plotrow}
\alias{check_plotrow}
\title{Keep or Remove Columns Specified by "plot"}
\usage{
check_plotrow(xs = NULL, sheet = NULL)
}
\arguments{
\item{xs}{"tbl_df" "tbl" "data.frame" excel spreadsheet read it from readxl::read_excel() with the first column title being \code{Samples}}

\item{sheet}{a character string provided by \code{readxl::excel_sheets()} for the excel file xs}
}
\value{
\itemize{
\item \code{check_plotrow} returns a tabular format similar to xs input \verb{"tbl_df" "tbl" "data.frame"}
\item it removes the final row xs\link{to_lower("plot"),}
}
}
\description{
Check if the final row contains optional "plot" filter & apply it if so
}
\examples{
data <- readxl::read_excel("www/flowjotter_example_data.xlsx", sheet = "CD4") |>
  dplyr::rename(`Samples` = 1) |>
  dplyr::select(dplyr::where(~ !all(is.na(.x)))) |>
  check_plotrow(xs = ., sheet = "CD4")
}
